+++ VKSample/VKSample/Models/Friend/Friend.swift
@@ -1,0 +1,9 @@
+// Friend.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Response about friends of user.
+struct Friend: Decodable {
+    let response: ResponseFriend
+}
