+++ VKSample/VKSample/Models/Friend/ResponseFriend.swift
@@ -1,0 +1,10 @@
+// ResponseFriend.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Friends list.
+struct ResponseFriend: Decodable {
+    let count: Int
+    let items: [ItemFriend]
+}
