+++ VKSample/VKSample/Models/Group/Group.swift
@@ -1,0 +1,9 @@
+// Group.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Response about groups.
+struct Group: Decodable {
+    let response: ResponseGroup
+}
