+++ VKSample/VKSample/Models/Group/ItemGroup.swift
@@ -1,19 +1,8 @@
-// Group.swift
+// ItemGroup.swift
 // Copyright © RoadMap. All rights reserved.
 
 import Foundation
 import RealmSwift
-
-/// Response about groups.
-struct Group: Decodable {
-    let response: ResponseGroup
-}
-
-/// Groups.
-struct ResponseGroup: Decodable {
-    let count: Int
-    let items: [ItemGroup]
-}
 
 /// Groups the user is subscribed to.
 final class ItemGroup: Object, Decodable {
