+++ VKSample/VKSample/Models/Group/ResponseGroup.swift
@@ -1,0 +1,10 @@
+// ResponseGroup.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Groups.
+struct ResponseGroup: Decodable {
+    let count: Int
+    let items: [ItemGroup]
+}
