+++ VKSample/VKSample/Models/Photo/Photo.swift
@@ -1,0 +1,9 @@
+// Photo.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Response about of user's photo.
+struct Photo: Decodable {
+    let response: ResponsePhoto
+}
