+++ VKSample/VKSample/Models/Photo/ResponsePhoto.swift
@@ -1,0 +1,10 @@
+// ResponsePhoto.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// User's photos.
+struct ResponsePhoto: Decodable {
+    let count: Int
+    let items: [ItemPhoto]
+}
